split textinput stories
